moved eventListener to utils
