+++ src/properties/pseudoClasses/userAction.js
@@ -1,5 +1,5 @@
 import State from '../../api/State'
-import createListener from '../../api/Listener'
+import createListener from '../../utils/createListener'
 
 const defaultKey = 'root'
